ユーザーに紐づくレビュー機能を追加

+++ src/lib/actions.ts
@@ -7,7 +7,6 @@ import { getBookById } from './getter';
 const prisma = new PrismaClient();
 
 export async function addReview(data: FormData) {
-    const book = await getBookById(data.get('id') as string);
 
     const email = data.get('email') as string;
     const user = await prisma.user.findUnique({
@@ -15,14 +14,31 @@ export async function addReview(data: FormData) {
         select: { id: true }
     });
 
+    if( !user ) {
+        throw new Error('User not found');
+    }
 
-    const input = {
+    const bookId = data.get('id') as string;
+    console.log('Adding review for bookId:', bookId);
+    // 既に本が存在するかチェック
+    const exists = await prisma.book.findUnique({ where: { id: bookId } });
+
+    // なければGoogle Books APIから取得して作成
+    if (!exists) {
+      const book = await getBookById(bookId);
+      await createBook({
+        id: book.id,
         title: book.title,
         author: book.author,
+        image: book.image ?? null,
         price: Number(book.price),
         publisher: book.publisher,
         published: book.published,
-        image: book.image,
+      });
+    }
+
+    const input = {
+        bookId: bookId,
         read: new Date(data.get('read') as string),
         memo: data.get('memo') as string,
         userId: user?.id || '',
@@ -30,20 +46,64 @@ export async function addReview(data: FormData) {
 
     await prisma.review.upsert({
         update: input,
-        create: { ...input, id: data.get('id') as string },
+        create: { ...input },
         where: {
-            id: data.get('id') as string
+            userId_bookId: {
+            userId: user.id,
+            bookId: bookId
+            }
         }
     });
 
     redirect('/');
 }
 
-export async function removeReview(id: string) {
+export async function removeReview(id: string, email: string) {
+    const user = await prisma.user.findUnique({
+        where: { email: email },
+        select: { id: true }
+    });
+    if (!user) {
+        throw new Error('User not found');
+    }
     await prisma.review.delete({
         where: {
-            id: id
+            userId_bookId: {
+                userId: user.id,
+                bookId: id
+            }
         }
     });
     redirect('/');
+}
+
+async function createBook(book: {
+  id: string;
+  title: string;
+  author: string;
+  image: string | null;
+  price: number;
+  publisher: string;
+  published: string;
+}) {
+  await prisma.book.upsert({
+    where: { id: book.id },
+    update: {
+      title: book.title,
+      author: book.author,
+      image: book.image,
+      price: book.price,
+      publisher: book.publisher,
+      published: book.published,
+    },
+    create: {
+      id: book.id,
+      title: book.title,
+      author: book.author,
+      image: book.image,
+      price: book.price,
+      publisher: book.publisher,
+      published: book.published,
+    },
+  });
 }
